requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.2.1
python-decouple==3.1
SQLAlchemy==1.3.10
Werkzeug==0.16.0
